generalize TermsOfUseNavigationPanel
